Fix potential overflow in segment_tree_mnnode

+++ code/data-structures/segment_tree_mnnode.cpp
@@ -2,13 +2,13 @@
 #define STNODE
 struct node {
   int l, r;
-  int x, lazy;
+  ll x, lazy;
   node() {}
   node(int _l, int _r) : l(_l), r(_r), x(INF), lazy(0) { }
-  node(int _l, int _r, int _x) : node(_l,_r) { x = _x; }
+  node(int _l, int _r, ll _x) : node(_l,_r) { x = _x; }
   node(node a, node b) : node(a.l,b.r) { x = min(a.x, b.x); }
-  void update(int v) { x = v; }
-  void range_update(int v) { lazy = v; }
+  void update(ll v) { x = v; }
+  void range_update(ll v) { lazy = v; }
   void apply() { x += lazy; lazy = 0; }
   void push(node &u) { u.lazy += lazy; } };
 #endif
